add check for link deleting

+++ src/main/java/com/dfedorino/urlshortener/ui/console/command/impl/DeleteLink.java
@@ -1,9 +1,11 @@
 package com.dfedorino.urlshortener.ui.console.command.impl;
 
 import com.dfedorino.urlshortener.domain.model.link.LinkDto;
+import com.dfedorino.urlshortener.domain.model.link.LinkStatus;
 import com.dfedorino.urlshortener.domain.model.user.User;
 import com.dfedorino.urlshortener.service.business.LinkService;
 import com.dfedorino.urlshortener.service.business.UserService;
+import com.dfedorino.urlshortener.service.validation.LinkValidationService.ValidatedLink;
 import com.dfedorino.urlshortener.ui.console.Cli;
 import com.dfedorino.urlshortener.ui.console.command.Command;
 import com.dfedorino.urlshortener.ui.console.command.dto.ResultWithNotification;
@@ -20,6 +22,7 @@ public class DeleteLink implements Command<LinkDto> {
     public static final String EXAMPLE_MESSAGE = KEY_TOKEN + " " + Cli.SHORTENED_URL_PREFIX + "AQ";
     public static final String SUCCESS_MESSAGE = "Link successfully deleted";
     public static final String DESCRIPTION_MESSAGE = "Delete given link";
+    public static final String LINK_ALREADY_DELETED = "Link %s already deleted";
     private final UserService userService;
     private final LinkService linkService;
 
@@ -43,13 +46,21 @@ public class DeleteLink implements Command<LinkDto> {
                     Command.USER_NOT_FOUND_MESSAGE.formatted(Cli.USER_UUID.get()));
         }
 
-        Optional<LinkDto> deleted = linkService.deleteLink(user.get().getId(), linkCode);
+        Optional<ValidatedLink> toBeDeleted = linkService.findValidatedLink(
+                user.get().getId(), linkCode);
 
-        if (deleted.isEmpty()) {
+        if (toBeDeleted.isEmpty()) {
             return ResultWithNotification.ofErrorMessage(
                     Command.SHORT_LINK_NOT_FOUND_MESSAGE.formatted(commandAndArgs[1]));
         }
 
+        ValidatedLink validatedLink = toBeDeleted.get();
+        if (LinkStatus.DELETED.name().equals(validatedLink.link().status())) {
+            return ResultWithNotification.ofErrorMessage(
+                    LINK_ALREADY_DELETED.formatted(commandAndArgs[1]));
+        }
+
+        Optional<LinkDto> deleted = linkService.deleteLink(user.get().getId(), linkCode);
         return new ResultWithNotification<>(SUCCESS_MESSAGE, deleted);
     }
